requirements.txt
loguru
mkdocs
pip
pytest
python-dotenv
ruff
tqdm
typer
qdrant-client
sentence-transformers
pandas
datasets
-e .
fastapi
uvicorn
python-multipart
google-generativeai
pillow
